refactor: change to CommonJs

+++ src/services/discordService.js
@@ -1,7 +1,7 @@
-import { Client, GatewayIntentBits } from "discord.js";
-import { WebhookClient } from "../WebhookClient.js";
-import MessageBuilder from "../builders/MessageBuilder.js";
-import EmbedBuilder from "../builders/EmbedBuilder.js";
+const { Client, GatewayIntentBits } = require("discord.js");
+const { WebhookClient } = require("../WebhookClient");
+const MessageBuilder = require("../builders/MessageBuilder");
+const EmbedBuilder = require("../builders/EmbedBuilder");
 
 let dmClient = null;
 
@@ -17,34 +17,23 @@ function buildNewsEmbed(newsContent) {
   embed.setTitle(`📰 ${newsContent.Title}`);
   if (newsContent.url) embed.setUrl(newsContent.url);
   embed.setDescription(newsContent.Passage);
-  if (newsContent.Category?.length) {
-    embed.addField("Category", newsContent.Category.join(", "), true);
-  }
-  if (newsContent.wordCount) {
-    embed.addField("Word Count", `${newsContent.wordCount} words`, true);
-  }
+  if (newsContent.Category?.length) embed.addField("Category", newsContent.Category.join(", "), true);
+  if (newsContent.wordCount) embed.addField("Word Count", `${newsContent.wordCount} words`, true);
   embed.setColor("#5865F2");
   return embed;
 }
 
-export async function sendReminder(vocabEmbeds, newsContent) {
+async function sendReminder(vocabEmbeds, newsContent) {
   const mode = process.env.MODE ?? "DM";
-  if (mode === "DM") {
-    await sendDm(vocabEmbeds, newsContent);
-  } else {
-    await sendWebhook(vocabEmbeds, newsContent);
-  }
+  if (mode === "DM") await sendDm(vocabEmbeds, newsContent);
+  else await sendWebhook(vocabEmbeds, newsContent);
 }
 
 async function sendDm(vocabEmbeds, newsContent) {
   const client = await getDmClient();
   const user = await client.users.fetch(process.env.DISCORD_ID);
   const dm = await user.createDM();
-
-  const vocabMessage = await dm.send({
-    embeds: vocabEmbeds.map((e) => e.toJSON()),
-  });
-
+  const vocabMessage = await dm.send({ embeds: vocabEmbeds.map((e) => e.toJSON()) });
   if (newsContent) {
     const newsEmbed = buildNewsEmbed(newsContent);
     await vocabMessage.reply({ embeds: [newsEmbed.toJSON()] });
@@ -53,20 +42,17 @@ async function sendDm(vocabEmbeds, newsContent) {
 
 async function sendWebhook(vocabEmbeds, newsContent) {
   const webhookClient = new WebhookClient(process.env.DISCORD_WEBHOOK);
-
   for (const embed of vocabEmbeds) {
     await webhookClient.send(new MessageBuilder().addEmbed(embed));
   }
-
   if (newsContent) {
     const newsEmbed = buildNewsEmbed(newsContent);
     await webhookClient.send(new MessageBuilder().addEmbed(newsEmbed));
   }
 }
 
-export async function destroyDmClient() {
-  if (dmClient) {
-    await dmClient.destroy();
-    dmClient = null;
-  }
+async function destroyDmClient() {
+  if (dmClient) { await dmClient.destroy(); dmClient = null; }
 }
+
+module.exports = { sendReminder, destroyDmClient };
